Simplify logic for saving and loading xia2 options
Serialise saved options to a JSON file.  Options to be saved are
determined by searching all the attributes of a UIOptionsWindow
instance.  If the type of an attribute matches a list of recognised
user-input widget types, then we record the value of a specific
pre-determined property of each widget.  For now, the recognised
user-input widget types are QCheckBox, QComboBox and QLineEdit, with the
corresponding properties 'checked', 'currentIndex' and 'text' recorded.